#78 nextjs 내 CSR/SSR 페이지에 따른 Service 객체 분리
- SSR 에서의 cookie 삽입과 CSR에서의 cookie 삽입 과정이 다르기에 해당 Service 메소드를 분리
    - CSR에서는 fetch API 내 credentials = include 를 통해 cookie 삽입 가능
    - SSR에서는 credential 이 동작하지 않으므로 nextjs에서 제공하는 headers()로 별도 주입해야함
