+++ frontend/src/service/ArticleSearchService.ts
@@ -1,4 +1,4 @@
-import { ArticleCreateRequest } from "@/interface";
+import {cookies} from "next/headers";
 
 const baseUrl = "http://localhost:8080/api/v2";
 
@@ -12,24 +12,13 @@ export async function getArticlePage(page: number, searchType: string, searchVal
 
 export async function getArticleDetail(articleId: number) {
     const response = await fetch(`${baseUrl}/articles/${articleId}`, {
+        method: 'GET',
         cache: 'no-store',
-        credentials: "include",
         headers: {
             "Content-Type": "application/json",
+            "Cookie": cookies().toString(),
         },
     });
 
     return response.json();
-}
-
-export async function createArticle(createRequest : ArticleCreateRequest) {
-    await fetch(`${baseUrl}/articles`, {
-        method: "POST",
-        credentials: "include",
-        cache: 'no-store',
-        headers: {
-            "Content-Type": "application/json",
-        },
-        body: JSON.stringify(createRequest),
-    });
 }
